requirements.txt
openenv>=0.1.0
numpy>=1.26.0
torch>=2.3.0
transformers>=4.44.0
unsloth[colab]>=2024.9.0
trl>=0.11.0
peft>=0.12.0
bitsandbytes>=0.43.0
accelerate>=0.33.0
datasets>=2.20.0
huggingface_hub>=0.24.0
spaces>=0.28.0
gradio>=4.42.0
matplotlib>=3.9.0
wandb>=0.17.0
tqdm>=4.66.0
pyyaml>=6.0
python-dotenv>=1.0.1
